Forward Input Events to GVRViews
The sample forwards key events generated by Android
to a GVRView.

Make use of the Gaze Controller to synthesize motion events
for GVRViews - use the cursor to click on buttons.

The sample also prints the headset hardware button presses
like back, volume up and down.

The sample only works with a VR headset.