added cusor and edittext.
added cusor and edittext.  the accuracy on picking the 'update' button
in the edittext isn't very good, but will work if you focus on the 'up'
portion of the button.  This should bring the repository up to date with
what I will present at SDC.

Also, DisplayUtils and the CustomPostEffectShaderManager looked to be in
their original directory.  was that deliberate?  or were they forgotten?
I moved them just in case.