Fix gvr-scripting demo onStep
The script bundle is attached really early in onCreate and it can't find
the "text" scene object when it tries to attach the bundle. I changed
one of the scripts to attach "text.js" to the "text" scene object
programatically instead of in the bundle. Now the onSensorEvent and
onStep functions get called.