Updated shadow mapping demo
To get shadow mapping to work correctly the application needs to set the near and far planes for the shadow map cameras using GVRLightBase.getShadowMaterial. I added this to the application. I did not move any of the objects, but I did change the direction of the spot and direct lights.

    changed light direction to point at objects to shadow
    removed commented out code
    enlarged spotlight cone angle
    fixed onStep logic to move light horizontally
    added touch event to start/stop light rotation
    use asset loader calls instead of deprecated GVRContext versions
    changed shadow light from aqua to white color
    remove unused point light