Fixed demos to work with latest controller fixes (#556)
* gvr-3dcursor: fix disappearing cursor

cursor disappears after settings dialog because cursor.setEnable(true) won't call onEnable if it already is enabled. I fixed this.

* revert last fix

* Fix gv4-3dcursor selection buttons

* Change demos to use GVRTouchPadGestureListener

Demos had several different implementations of
VRTouchPadGestureDetector. We cannot legally distribute this class in
the Framework. Demos have been changed to use the new
GVRTouchPadGestureListener class which does pretty much the same thing
but in a different way. Now all these demos use the same class for this
functionality.

* Use GVREventReceiver for controller event dispatch

* Remove CursorUtils.java from gvr-remote-scripting

* gvr-accessibility: dont add cursor to scene

Let the input manager do it

* Make sendEventsToActivity the default

* gvr-widgetviewer: clean up the code

1. use GVRSwitch object to select between objects
2. isolate unused thumbnail code
3. refactor into multiple functions

* Daydream build for controller-ready demos

changed AndroidManifest.xml to include DAYDREAM intent for the apps which support the daydream controller

* Removed buildToolsVersion from build.gradle files

* Remove useControllerTypes from gvr.xml

To test the default of "gaze,controller" which is what these demos require

* fix reticle in gvr-controller demo